Upgrade to react 19 and react compiler Also cleanup some code, add nice transitions, make pixelate look nicer
