+++ .prettierignore
@@ -1,3 +1,4 @@
 coverage/
 dist/
 pnpm-lock.yaml
+.DS_Store
