+++ package.json
@@ -12,28 +12,31 @@
     "pnpm": ">=7.1.0"
   },
   "dependencies": {
-    "@react-spring/web": "^9.5.5",
+    "@react-spring/web": "^9.7.4",
     "colord": "^2.9.3",
     "eslint-plugin-react-hooks": "^4.6.0",
     "fast-simplex-noise": "^4.0.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
+    "@babel/plugin-transform-react-jsx-development": "^7.25.9",
     "@swc/core": "^1.3.3",
     "@trivago/prettier-plugin-sort-imports": "^3.3.0",
     "@types/node": "^18.7.20",
     "@types/prettier": "^2.7.1",
-    "@types/react": "^18.0.21",
-    "@types/react-dom": "^18.0.6",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
     "@typescript-eslint/eslint-plugin": "^5.38.0",
     "@typescript-eslint/parser": "^5.38.0",
-    "@vitejs/plugin-react": "^2.1.0",
+    "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.12",
+    "babel-plugin-react-compiler": "19.0.0-beta-df7b47d-20241124",
     "eslint": "^8.24.0",
     "eslint-import-resolver-typescript": "^3.5.1",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-react": "^7.31.8",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-compiler": "19.0.0-beta-df7b47d-20241124",
     "eslint-plugin-typescript-sort-keys": "^2.1.0",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.16",
@@ -42,7 +45,7 @@
     "tailwindcss": "^3.1.8",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.3",
-    "vite": "^3.1.3",
+    "vite": "^6.0.3",
     "vitest": "^0.23.4"
   },
   "scripts": {
