+++ src/App.css
@@ -2,10 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 
-.filter {
-  filter: url('#turb');
-}
-
 .pixelate {
   filter: url('#pixelate');
 }
