+++ src/constants.ts
@@ -1,0 +1,17 @@
+export const colorsNatural = [
+  'hsla(330 100% 80%)',
+  'hsla(280 100% 90%)',
+  'hsla(250 10% 100%)',
+  'hsla(170 60% 40%)',
+  'hsla(110 45% 60%)',
+  'hsla(90 65% 70%)',
+  'hsla(60 75% 75%)',
+  'hsla(200 80% 65%)',
+  '#008fb0',
+  '#0c4278',
+].reverse();
+
+export const mapSize = 120;
+export const levels = 7;
+export const baseTileSize = 70;
+export const perimeter = 12;
